Split sdk and ffi jars
