+++ ffi/src/main/java/com/antithesis/ffi/internal/OutputHandler.java
@@ -1,0 +1,12 @@
+package com.antithesis.ffi.internal;
+
+public interface OutputHandler {
+    void output(String value);
+
+    long random();
+
+    long initializeModuleCoverage(long edgeCount, String symbolFilePath);
+
+    void notifyModuleEdge(long edgePlusModule);
+}
+
