+++ practice.properties
@@ -1,0 +1,3 @@
+org.gradle.java.installations.auto-download=false
+version = 1.3.0-test
+
