+++ sdk/src/main/java/com/antithesis/sdk/internal/LocationInfo.java
