--- src/main/java/com/antithesis/sdk/internal/OutputHandler.java
@@ -1,11 +1,0 @@
-package com.antithesis.sdk.internal;
-
-public interface OutputHandler {
-    void output(String value);
-
-    long random();
-
-    long initializeModuleCoverage(long edgeCount, String symbolFilePath);
-
-    void notifyModuleEdge(long edgePlusModule);
-}
